chore: Split task and task info
